[#15] ✨ 회원 목록 조회 화면 구현 (#16)

+++ domain/src/test/kotlin/com/splab/springgames/domain/member/vo/GameCardTotalPriceTest.kt
@@ -15,7 +15,7 @@ class GameCardTotalPriceTest : DescribeSpec({
                 val value = 0
 
                 // act
-                val gameCardTotalPrice = GameCardTotalPrice(value)
+                val gameCardTotalPrice = GameCardTotalPrice(value.toBigDecimal())
 
                 // assert
                 gameCardTotalPrice.value shouldBe value
@@ -29,7 +29,7 @@ class GameCardTotalPriceTest : DescribeSpec({
 
                 // act & assert
                 shouldThrow<IllegalArgumentException> {
-                    GameCardTotalPrice(value)
+                    GameCardTotalPrice(value.toBigDecimal())
                 }
             }
         }
